Changed: Merge catalog.rs into mod.rs and improve encapsulation
Consolidate ModelCatalog into the module root and use a constructor
instead of pub(crate) fields.

Changes:
- Merged catalog.rs into mod.rs for simpler module structure
- Added ModelCatalog::new() constructor, made fields private
- Builder now uses constructor instead of struct literal
- Cleaned up unused re-exports in public/mod.rs

Benefits:
- Better encapsulation with private fields
- Flatter module hierarchy
- Cleaner public API surface
