+++ src/llm-coding-tools-core/src/models/catalog/public/mod.rs
@@ -3,10 +3,7 @@
 //! This module contains public-facing data types used when building
 //! and querying a [`ModelCatalog`].
 
-pub use builder_types::{
-    LookupTableKind, ModelCatalogBuildError, ModelInfo, ModelSourceRow, ProviderInfo,
-    ProviderSourceRow,
-};
+pub use builder_types::{LookupTableKind, ModelInfo, ProviderInfo};
 pub use entry::{CatalogEntry, Model, Provider};
 pub use modality::Modality;
 pub use model_idx::ModelIdx;
